refactor: split Elo computation into modules
